requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10.5
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-taggit==0.22.1
django-taggit-templatetags2==1.6.1
gunicorn==19.7.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
whitenoise==3.3.0
